+++ src/message_manager.py
@@ -1,28 +1,30 @@
 import time
 import datetime
 import os
-import json
+import logging
 from access_manager import AccessManager
 from chat_session import ChatSession
 from openai_parser import OpenAIParser
+from config_loader import ConfigLoader
 
 
 class MessageManager:
 
     userDict = {}
-    config_dict = {}
+    # config_dict = {}
     openai_parser = None
     access_manager = None
     
     def __init__(self, access_manager):
         self.openai_parser = OpenAIParser()
         self.access_manager = access_manager
-        # load config
-        with open("config.json") as f:
-            self.config_dict = json.load(f)
+        # setup logger
+        self.logger = logging.getLogger("MessageManager")
+        
+        self.access_manager = AccessManager()
 
     def get_response(self, id, user, message):
-
+        self.logger.debug("Get response for user: %s" % id)
         t = time.time()
 
         # (permission, clue) = self.access_manager.check_user_allowed(user)
@@ -43,20 +45,18 @@ class MessageManager:
         return answer
 
     def clear_context(self, id):
+        self.logger.debug("Clear context for user: %s" % id)
         try:
             self.userDict[id].clear_context(time.time())
         except Exception as e:
             print(e)
 
     def get_generated_image_url(self, user, prompt, num=1):
+        self.logger.debug("Get generated image for user: %s" % user)
 
-        # Temporary fix by @Flynn, will be fixed in the next version
-        with open("config.json") as f:
-            super_users = json.load(f)["super_users"]
-        if user in super_users:
+        if user in ConfigLoader.get("super_users"):
             url, _ = self.openai_parser.image_generation(user, prompt)
-            return url, "Hey boss, it's on your account. 💰"
-        ############################
+            return (url, "Hey boss, it's on your account. 💰")
 
         (permission, clue) = self.access_manager.check_image_generation_allowed(user, num)
         if permission == False:
@@ -68,15 +68,27 @@ class MessageManager:
         return url, clue
 
     def get_transcript(self, user, audio_file):
+        self.logger.debug("Get voice transcript for user: %s" % user)
         # (permission, clue) = self.access_manager.check_user_allowed(user)
         # if permission == False:
         #     return clue
 
-        try:
-            return self.openai_parser.speech_to_text(user, audio_file)
-        except Exception as e:
-            print(e)
-            return ""
+        return self.openai_parser.speech_to_text(user, audio_file)
+    
+    def set_system_role(self, id, user, message):
+        self.logger.debug("Set system role for chat: %s" % id)
+        t = time.time()
+        if id not in self.userDict:
+            self.userDict[id] = ChatSession(t, message)       
+        self.userDict[id].set_system_role(t, message)   
+        
+        # send first sentence
+        (answer, usage) = self.__sendMessage(user, 
+                [{"role": "system", "content": message}, 
+                 {"role": "user", "content":"Say hello to me."}])
+        self.access_manager.update_usage_info(user, usage, "chat")
+        return answer
+        
 
     def __sendMessage(self, user, messageList):
         ans = self.openai_parser.get_response(user, messageList)